pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cairn-memory"
version = "0.1.0"
description = "Cairn — a persistent reasoning graph that keeps humans and agents oriented across sessions"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "numpy>=1.24,<3",
    "mcp[cli]>=1.0",
    "networkx>=3.0,<4",
    "pydantic>=2.0,<3",
    "python-dotenv>=1.0",
    "voyageai>=0.3.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cairn"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require ANTHROPIC_API_KEY + VOYAGE_API_KEY)",
]
